requirements.txt
ViennaRNA
tqdm